# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/Login/{Username}/{Password}": {
            "get": {
                "description": "this is a request to give the middleware user information. this will allow the middleware to set up the authentication token need to start and stop the recording.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Send middleware user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": ":Username",
                        "name": "Username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": ":Password",
                        "name": "Password",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/Start/{Usecase}/{Appiden}": {
            "get": {
                "description": "This endpoint is to stop a recording and needs a usecase and a applicationIdentifier as parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Start a recording",
                "parameters": [
                    {
                        "type": "string",
                        "description": ":Usecase",
                        "name": "Usecase",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": ":Appiden",
                        "name": "Appiden",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/Stop/{Usecase}/{Appiden}": {
            "get": {
                "description": "This endpoint is to stop a recording and needs a usecase and a applicationIdentifier as parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "Stop a recording",
                "parameters": [
                    {
                        "type": "string",
                        "description": ":Usecase",
                        "name": "Usecase",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": ":Appiden",
                        "name": "Appiden",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    }
}